Utilities: Improved hexdump() function, added ascii char column
Improved hexdump() function to print in a more familiar format,
including an additional column displaying the data as ascii chars.

Also use correct datatypes for variables so that the function can now
display data buffers above 64kb.

+++ wic64/util.h
@@ -3,6 +3,6 @@
 
 #include <cstdint>
 
-void hexdump(const char *msg, uint8_t *data, uint16_t size);
+void hexdump(const char *msg, uint8_t *data, uint32_t size);
 
 #endif // WIC64_HEXDUMP_H
